Rename solution projects and change namespaces for the projects
