+++ src/Casparcg.Core.Device/CGXmlData.cs
@@ -1,0 +1,36 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Casparcg.Core.Device
+{
+	public class CGXmlData : ICGComponentData
+    {
+		public CGXmlData(string xmlString)
+		{
+			XmlString = xmlString;
+		}
+
+		public string XmlString { get; set; }
+		#region ICGComponentData Members
+
+		public void ToXml(StringBuilder sb)
+		{
+			sb.Append("<data id=\"xml\">");
+			sb.Append(XmlString);
+			sb.Append("</data>");
+		}
+
+		public void ToAMCPEscapedXml(StringBuilder sb)
+		{
+			sb.Append("<data id=\\\"xml\\\">");
+			string escapedXml = XmlString.Replace("\\", "\\\\");
+			escapedXml = escapedXml.Replace("\"", "\\\"");
+			sb.Append(escapedXml);
+			sb.Append("</data>");
+		}
+
+
+		#endregion
+	}
+}
