+++ src/Casparcg.Core.Device/Easing.cs
@@ -1,0 +1,54 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace Casparcg.Core.Device
+{
+    public enum Easing
+    {
+        None,
+        Linear,
+        EaseNone,
+        EaseInQuad,
+        EaseOutQuad,
+        EaseInOutQuad,
+        EaseOutInQuad,
+        EaseInCubic,
+        EaseOutCubic,
+        EaseInOutCubic,
+        EaseOutInCubic,
+        EaseInQuart,
+        EaseOutQuart,
+        EaseInOutQuart,
+        EaseOutInQuart,
+        EaseInQuint,
+        EaseOutQuint,
+        EaseInOutQuint,
+        EaseOutInQuint,
+        EaseInSine,
+        EaseOutSine,
+        EaseInOutSine,
+        EaseOutInSine,
+        EaseInExpo,
+        EaseOutExpo,
+        EaseInOutExpo,
+        EaseOutInExpo,
+        EaseInCirc,
+        EaseOutCirc,
+        EaseInOutCirc,
+        EaseOutInCirc,
+        EaseInElastic,
+        EaseOutElastic,
+        EaseInOutElastic,
+        EaseOutInElastic,
+        EaseInBack,
+        EaseOutBack,
+        EaseInOutBack,
+        EaseOutInBack,
+        EaseOutBounce,
+        EaseInBounce,
+        EaseInOutBounce,
+        EaseOutInBounce
+    }
+}
