+++ src/Casparcg.Core.Device/Properties/Resources.resx
