+++ src/Casparcg.Core.Device/Properties/Settings.settings
