+++ src/Casparcg.Core.Device/app.config
