+++ src/Casparcg.Core.Network/ProtocolStrategy.cs
@@ -1,0 +1,19 @@
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Casparcg.Core.Network
+{
+	public interface IProtocolStrategy
+	{
+        void Parse(string str, RemoteHostState state);
+        void Parse(byte[] data, int length, RemoteHostState state);
+
+		System.Text.Encoding Encoding {
+			get;
+		}
+		string Delimiter {
+			get;
+		}
+	}
+}
