+++ src/Solution.sln
@@ -1,17 +1,13 @@
 ﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Svt.Caspar", "Svt.Caspar\Svt.Caspar.csproj", "{9B56A150-ECE3-4099-AEA1-0F8E115DC374}"
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27130.2024
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Casparcg.Core.Device", "Casparcg.Core.Device\Casparcg.Core.Device.csproj", "{9B56A150-ECE3-4099-AEA1-0F8E115DC374}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Svt.Network", "Svt.Network\Svt.Network.csproj", "{941ED6A5-C8DF-459D-BFCE-5C6ED9AC0492}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Svt.Caspar.Controls", "Svt.Caspar.Controls\Svt.Caspar.Controls.csproj", "{D5CEE141-6C59-4837-A576-21BA098CC813}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CasparCG.Core.Network", "Casparcg.Core.Network\CasparCG.Core.Network.csproj", "{941ED6A5-C8DF-459D-BFCE-5C6ED9AC0492}"
 EndProject
 Global
-	GlobalSection(SubversionScc) = preSolution
-		Svn-Managed = True
-		Manager = AnkhSVN - Subversion Support for Visual Studio
-	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Debug|Mixed Platforms = Debug|Mixed Platforms
@@ -105,41 +101,15 @@ Global
 		{941ED6A5-C8DF-459D-BFCE-5C6ED9AC0492}.Unit test|x64.ActiveCfg = Develop|x64
 		{941ED6A5-C8DF-459D-BFCE-5C6ED9AC0492}.Unit test|x64.Build.0 = Develop|x64
 		{941ED6A5-C8DF-459D-BFCE-5C6ED9AC0492}.Unit test|x86.ActiveCfg = Develop|x64
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Debug|Mixed Platforms.ActiveCfg = Debug|x64
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Debug|Mixed Platforms.Build.0 = Debug|x64
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Debug|Win32.ActiveCfg = Debug|x64
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Debug|x64.ActiveCfg = Debug|x64
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Debug|x64.Build.0 = Debug|x64
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Debug|x86.ActiveCfg = Debug|x86
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Debug|x86.Build.0 = Debug|x86
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Develop|Any CPU.ActiveCfg = Develop|Any CPU
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Develop|Any CPU.Build.0 = Develop|Any CPU
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Develop|Mixed Platforms.ActiveCfg = Develop|x64
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Develop|Mixed Platforms.Build.0 = Develop|x64
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Develop|Win32.ActiveCfg = Develop|x64
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Develop|x64.ActiveCfg = Develop|x64
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Develop|x64.Build.0 = Develop|x64
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Develop|x86.ActiveCfg = Develop|x86
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Develop|x86.Build.0 = Develop|x86
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Release|Mixed Platforms.ActiveCfg = Release|x64
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Release|Mixed Platforms.Build.0 = Release|x64
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Release|Win32.ActiveCfg = Release|x64
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Release|x64.ActiveCfg = Release|x64
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Release|x64.Build.0 = Release|x64
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Release|x86.ActiveCfg = Release|x86
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Release|x86.Build.0 = Release|x86
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Unit test|Any CPU.ActiveCfg = Develop|x64
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Unit test|Mixed Platforms.ActiveCfg = Develop|x64
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Unit test|Mixed Platforms.Build.0 = Develop|x64
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Unit test|Win32.ActiveCfg = Develop|x64
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Unit test|x64.ActiveCfg = Develop|x64
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Unit test|x64.Build.0 = Develop|x64
-		{D5CEE141-6C59-4837-A576-21BA098CC813}.Unit test|x86.ActiveCfg = Develop|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {80DA2F89-8337-4566-8BA0-40B40AD40DCA}
+	EndGlobalSection
+	GlobalSection(SubversionScc) = preSolution
+		Svn-Managed = True
+		Manager = AnkhSVN - Subversion Support for Visual Studio
 	EndGlobalSection
 EndGlobal
