--- src/Svt.Caspar/Easing.cs
@@ -1,54 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Svt.Caspar
-{
-    public enum Easing
-    {
-        None,
-        Linear,
-        EaseNone,
-        EaseInQuad,
-        EaseOutQuad,
-        EaseInOutQuad,
-        EaseOutInQuad,
-        EaseInCubic,
-        EaseOutCubic,
-        EaseInOutCubic,
-        EaseOutInCubic,
-        EaseInQuart,
-        EaseOutQuart,
-        EaseInOutQuart,
-        EaseOutInQuart,
-        EaseInQuint,
-        EaseOutQuint,
-        EaseInOutQuint,
-        EaseOutInQuint,
-        EaseInSine,
-        EaseOutSine,
-        EaseInOutSine,
-        EaseOutInSine,
-        EaseInExpo,
-        EaseOutExpo,
-        EaseInOutExpo,
-        EaseOutInExpo,
-        EaseInCirc,
-        EaseOutCirc,
-        EaseInOutCirc,
-        EaseOutInCirc,
-        EaseInElastic,
-        EaseOutElastic,
-        EaseInOutElastic,
-        EaseOutInElastic,
-        EaseInBack,
-        EaseOutBack,
-        EaseInOutBack,
-        EaseOutInBack,
-        EaseOutBounce,
-        EaseInBounce,
-        EaseInOutBounce,
-        EaseOutInBounce
-    }
-}
